Complemted get message flow
