+++ client/src/main/resources/application.conf
@@ -48,3 +48,8 @@ db.config {
   search.limit = 10
   search.limit = ${?SEARCH_LIMIT}
 }
+
+scheduler {
+    initial.delay = 10
+    interval = 3
+}
